requirements.txt
google-cloud-aiplatform
google-adk
mcp
vertexai